--- vehicle-ads/PULL_REQUEST.md
@@ -1,5 +1,0 @@
-#Vehicle Ads
-
-Pensando no desenvolvimento do projeto, optei pela criação de somente uma pagina inicial, contendo todas as informações necessárias para a criação de novos veículos no banco de dados.
-
-Não foi possível o deploy da aplicação por erros no Vercel e no Netlify, entretanto, caso seja possível, até futura entrevista, imagino conseguir deployar corretamente.
