+++ vehicle-ads/src/App.jsx
@@ -1,5 +1,5 @@
 import { BrowserRouter } from "react-router-dom";
-import { Router } from "./routes/Router";
+import Router  from "./routes/Router";
 
 function App() {
   return (
