--- vehicle-ads/src/components/Central.tsx
@@ -1,8 +1,0 @@
-import { ReactNode } from "react";
-
-type ChildrenComponentProps = {
-  children: ReactNode;
-};
-export default function Central({ children }: ChildrenComponentProps) {
-  return <div className="bg-white flex flex-col">{children}</div>;
-}
