+++ vehicle-ads/src/routes/Router.jsx
@@ -1,0 +1,10 @@
+import { Routes, Route } from "react-router-dom";
+import Home  from "../pages/Home/Home";
+
+export default function Router() {
+    return (
+        <Routes>
+            <Route path="/" element={ <Home />} />
+        </Routes>
+    )
+}
