--- vehicle-ads/src/routes/Router.tsx
@@ -1,12 +1,0 @@
-import { Routes, Route } from "react-router-dom";
-
-import { Home } from "../pages/Home";
-
-export function Router() {
-  return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/search/:slug" element={<Home />} />
-    </Routes>
-  );
-}
