--- vehicle-ads/src/service/api.tsx
@@ -1,11 +1,0 @@
-import axios from "axios";
-
-const api = axios.create({
-  baseURL: "https://api-vehicleads.herokuapp.com",
-  timeout: 10000,
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-export default api;
