--- vehicle-ads/tailwind.config.js
@@ -1,8 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./src/**/*.tsx", "./src/**/*.jsx"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
