+++ vehicle-ads/vite.config.js
